Reduce compiler warnings (#2456)
* Exclude -fno-rtti option from compilation of c files

* Change string to char type for equal sign in android input

* Removed unused variables

* Remove double const declaration

* Move platform specific variable in same scope as usage

* Fix unitiallized/no return path warnings

* Fix type conversion warnings

* Use correct path errno.h include

* Explicitly ignore fread() return value

* Remove obsolete -Wno-warn-absolute-paths command line option for web build

* Fixed compiler warning for win32

* Explicitly use float versions of atan2 and assign

* More type conversion warning fixes in hashtable.h for win32

* Don’t define be64toh if it already exists (it does on Linux)

* Win32 no longer requires declaration of alloca()

* Moved dmGui::DEFAULT_LAYER and DEFAULT_LAYOUT value to cpp file

* Use uint32_t for active_attachment in rig

* Move dmGui::DEFAULT_LAYER inside of gui.cpp

* Review fixes

+++ engine/dlib/src/dmsdk/dlib/array.h
@@ -377,13 +377,13 @@ const T& dmArray<T>::Back() const
 template <typename T>
 uint32_t dmArray<T>::Size() const
 {
-    return (uint32_t)(m_End - m_Front);
+    return (uint32_t)(uintptr_t)(m_End - m_Front);
 }
 
 template <typename T>
 uint32_t dmArray<T>::Capacity() const
 {
-    return (uint32_t)(m_Back - m_Front);
+    return (uint32_t)(uintptr_t)(m_Back - m_Front);
 }
 
 template <typename T>
@@ -401,7 +401,7 @@ bool dmArray<T>::Empty() const
 template <typename T>
 uint32_t dmArray<T>::Remaining() const
 {
-    return m_Back - m_End;
+    return (uint32_t)(uintptr_t)(m_Back - m_End);
 }
 
 template <typename T>
